chore: move analytics stories to external file
