+++ dev/app/builtin/init-stories.js
@@ -1,33 +1,14 @@
 /* eslint-disable import/default */
 
-import { action, storiesOf } from 'dev-novel';
+import { storiesOf } from 'dev-novel';
 import instantsearch from '../../../index.js';
 import wrapWithHits from '../utils/wrap-with-hits.js';
 import initInstantSearchStories from './stories/instantsearch.stories';
+import initAnalyticsStories from './stories/analytics.stories';
 
 export default () => {
   initInstantSearchStories();
-
-  storiesOf('Analytics').add(
-    'default',
-    wrapWithHits(container => {
-      const description = document.createElement('p');
-      description.innerText = 'Search for something, look into Action Logger';
-      container.appendChild(description);
-
-      window.search.addWidget(
-        instantsearch.widgets.analytics({
-          pushFunction(formattedParameters, state, results) {
-            action('pushFunction[formattedParameters]')(formattedParameters);
-            action('pushFunction[state]')(state);
-            action('pushFunction[results]')(results);
-          },
-          triggerOnUIInteraction: true,
-          pushInitialSearch: false,
-        })
-      );
-    })
-  );
+  initAnalyticsStories();
 
   storiesOf('SearchBox')
     .add(
